Fix stencil to use uint8_t and allow clearing data on RenderGraphNode

+++ src/engine/core/include/halley/core/graphics/painter.h
@@ -42,7 +42,7 @@ namespace Halley
 		Camera& getCurrentCamera() const { return *camera; }
 		Rect4f getWorldViewAABB() const;
 
-		virtual void clear(std::optional<Colour> colour, std::optional<float> depth = 1.0f, std::optional<uint32_t> stencil = 0) = 0;
+		virtual void clear(std::optional<Colour> colour, std::optional<float> depth = 1.0f, std::optional<uint8_t> stencil = 0) = 0;
 		virtual void setMaterialPass(const Material& material, int pass) = 0;
 		virtual void setMaterialData(const Material& material) = 0;
